backend/requirements.txt
Django==4.2.7
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8 